Updates nav to be more mobile friendly
Updates the nav to change to a dropdown menu for display less than
768px. Also hides the large logo on the home page when displaying the
dropdown nav since it looked strange with the logo in 2 places.
It's not an ideal solution but it does address the issue with the menu
disappearing.

Also adds media query to allow the sub page logo to scale between 768px
and 860px wide viewports. There were some display issues, but it didn't
look right defaulting to the dropdown at large sizes.